Add VeluxEntity to abstract away common methods for all entities
