+++ homeassistant/components/velux/__init__.py
@@ -5,8 +5,10 @@ from pyvlx import PyVLX, PyVLXException
 import voluptuous as vol
 
 from homeassistant.const import CONF_HOST, CONF_PASSWORD, EVENT_HOMEASSISTANT_STOP
+from homeassistant.core import callback
 from homeassistant.helpers import discovery
 import homeassistant.helpers.config_validation as cv
+from homeassistant.helpers.entity import Entity
 
 DOMAIN = "velux"
 DATA_VELUX = "data_velux"
@@ -75,3 +77,42 @@ class VeluxModule:
         _LOGGER.debug("Velux interface started")
         await self.pyvlx.load_scenes()
         await self.pyvlx.load_nodes()
+
+
+class VeluxEntity(Entity):
+    """Abstraction for al Velux entities."""
+
+    def __init__(self, node):
+        """Initialize the Velux device."""
+        self.node = node
+
+    @callback
+    def async_register_callbacks(self):
+        """Register callbacks to update hass after device was changed."""
+
+        async def after_update_callback(device):
+            """Call after device was updated."""
+            self.async_write_ha_state()
+
+        self.node.register_device_updated_cb(after_update_callback)
+
+    async def async_added_to_hass(self):
+        """Store register state change callback."""
+        self.async_register_callbacks()
+
+    @property
+    def unique_id(self) -> str:
+        """Return the unique id base on the serial_id returned by Velux."""
+        return self.node.serial_number
+
+    @property
+    def name(self):
+        """Return the name of the Velux device."""
+        if not self.node.name:
+            return "#" + str(self.node.node_id)
+        return self.node.name
+
+    @property
+    def should_poll(self):
+        """No polling needed within Velux."""
+        return False
