+++ homeassistant/components/velux/light.py
@@ -6,9 +6,8 @@ from homeassistant.components.light import (
     SUPPORT_BRIGHTNESS,
     LightEntity,
 )
-from homeassistant.core import callback
 
-from . import DATA_VELUX
+from . import DATA_VELUX, VeluxEntity
 
 
 async def async_setup_platform(hass, config, async_add_entities, discovery_info=None):
@@ -20,38 +19,8 @@ async def async_setup_platform(hass, config, async_add_entities, discovery_info=
     async_add_entities(entities)
 
 
-class VeluxLight(LightEntity):
+class VeluxLight(VeluxEntity, LightEntity):
     """Representation of a Velux light."""
-
-    def __init__(self, node):
-        """Initialize the light."""
-        self.node = node
-
-    @callback
-    def async_register_callbacks(self):
-        """Register callbacks to update hass after device was changed."""
-
-        async def after_update_callback(device):
-            """Call after device was updated."""
-            self.async_write_ha_state()
-
-        self.node.register_device_updated_cb(after_update_callback)
-
-    async def async_added_to_hass(self):
-        """Store register state change callback."""
-        self.async_register_callbacks()
-
-    @property
-    def unique_id(self):
-        """Return the unique ID of this cover."""
-        return self.node.serial_number
-
-    @property
-    def name(self):
-        """Return the name of the Velux device."""
-        if not self.node.name:
-            return "Light #" + str(self.node.node_id)
-        return self.node.name
 
     @property
     def supported_features(self):
